Clean separation between backend and enrichment libs
