+++ enrich/crunchbase/crunchbase.js
@@ -4,7 +4,7 @@ const {
   getUrlDomain,
   getSocialPath,
   camelizeKeys,
-} = require("../../backend/src/utils")
+} = require("../utils")
 
 const API_KEY = process.env.CRUNCHBASE_API_KEY
 const ODM_ORGS_URL = "https://api.crunchbase.com/v3.1/odm-organizations"
